+++ src/directives/logistics/evacuateTerminal.ts
@@ -1,12 +1,12 @@
 import {profile} from '../../profiler/decorator';
 import {Directive} from '../Directive';
-import {log} from '../../lib/logger/log';
+import {log} from '../../console/log';
 
 @profile
-export class DirectiveAbandon extends Directive {
+export class DirectiveEvacuateTerminal extends Directive {
 
-	static directiveName = 'abandon';
-	static color = COLOR_PURPLE;
+	static directiveName = 'evacuateTerminal';
+	static color = COLOR_YELLOW;
 	static secondaryColor = COLOR_RED;
 
 	// colony: Colony | undefined; // this is technically unallowable, but at end of life, colony can be undefined
@@ -22,6 +22,9 @@ export class DirectiveAbandon extends Directive {
 		}
 		// Register abandon status
 		this.colony.abandoning = true;
+		if (Game.time % 25 == 0) {
+			log.alert(`${this.pos.print}: terminal evacuation in progress!`);
+		}
 	}
 
 	init() {