+++ src/tasks/index.d.ts
@@ -2,6 +2,8 @@ interface TaskSettings {
 	targetRange: number;
 	workOffRoad: boolean;
 	oneShot: boolean;
+	timeout: number;
+	blind: boolean;
 }
 
 interface TaskOptions {
@@ -31,47 +33,4 @@ interface protoTask {
 	tick: number;
 	options: TaskOptions;
 	data: TaskData;
-}
-
-interface ITask extends protoTask {
-	settings: TaskSettings;
-	proto: protoTask;
-	creep: Creep;
-	target: RoomObject | null;
-	targetPos: RoomPosition;
-	parent: ITask | null;
-	manifest: ITask[];
-	targetManifest: (RoomObject | null)[];
-	targetPosManifest: RoomPosition[];
-	eta: number | undefined;
-
-	fork(newTask: ITask): ITask;
-
-	isValidTask(): boolean;
-
-	isValidTarget(): boolean;
-
-	isValid(): boolean;
-
-	move(): number;
-
-	moveToNextPos(): number | undefined;
-
-	run(): number | void;
-
-	work(): number;
-
-	finish(): void;
-}
-
-// interface CreepMemory {
-// 	task: protoTask | null;
-// }
-
-interface Creep {
-	task: ITask | null;
-	hasValidTask: boolean;
-	isIdle: boolean;
-
-	run(): number | void;
 }