+++ src/utilities/WorldMap.ts
@@ -5,7 +5,7 @@ import {profile} from '../profiler/decorator';
 
 export const ROOMTYPE_SOURCEKEEPER = 'SK';
 export const ROOMTYPE_CORE = 'CORE';
-export const ROOMTYPE_CONTROLLER = 'CTRLR';
+export const ROOMTYPE_CONTROLLER = 'CTRL';
 export const ROOMTYPE_ALLEY = 'ALLEY';
 
 @profile
@@ -91,7 +91,6 @@ export class WorldMap {
 					return 6;
 				}
 			} else {
-				// must be the same missionRoom, no direction
 				return 0;
 			}
 		} else {
@@ -128,7 +127,6 @@ export class WorldMap {
 	}
 
 	public static getRoomCoordinates(roomName: string): RoomCoord {
-
 		let coordinateRegex = /(E|W)(\d+)(N|S)(\d+)/g;
 		let match = coordinateRegex.exec(roomName)!;